Fix a compiler warning.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@17296 3db68df2-c116-0410-a063-a993310a9797

+++ source/simulation2/components/CCmpRangeManager.cpp
@@ -736,7 +736,7 @@ public:
 			m_DirtyVisibility.clear();
 			m_DirtyVisibility.resize(m_LosTilesPerSide*m_LosTilesPerSide);
 		}
-		ENSURE(m_DirtyVisibility.size() == m_LosTilesPerSide*m_LosTilesPerSide);
+		ENSURE(m_DirtyVisibility.size() == (size_t)(m_LosTilesPerSide*m_LosTilesPerSide));
 
 		m_LosTiles.clear();
 		m_LosTiles.resize(m_LosTilesPerSide*m_LosTilesPerSide);
